ec: fix representation of point at infinity #94
Affine coordinates cannot represent the point at infinity. Technically,
it's possible that (0,0) (the previous shorthand for pai) could be on
the curve, which would make this implementation incorrect. In practice,
it's not on any of the points we care about, but it sketches me out to
have a haphazard implementation like that.