ecdsa: separate parameter-size constraints #94
After some offline discussion, we decided it would be nice to have a
separate specification that enforces the constraints on hash and curve
size that don't affect the computation of the signature. Now the
implementation is divided into two versions, one that enforces those
constraints and one that doesn't.

Also changes example file to be for a valid curve/hash combo.