modes: CBC implementation based on CipherInterface #80
The previous implementation of CBC (not yet removed) takes as input a
keysize and an encrypt/decrypt function, and hardcodes the IV and block
size to 128 bits.

Instead, we'd like to make use of CipherInterface, which makes it easier
to instantiate CBC with different block ciphers (and without any
hardcoded assumptions on the block size).

Removal of the old implementation and updating module-level comments is
reserved for future commits