wots: copy public API into an interface #175
The practical use case for WOTS+ is as a component of other protocols.
These are typically implemented generically over WOTS+, and then
instantiated with a joint set of parameters for WOTS+ and the parent
scheme.

This means that just instantiating WOTS+ with the approved parameter
sets isn't suitable, since we want to define other protocols based on
the generic specification.

This commit adds an interface that defines the parameters (n, w, F, PRF)
and the public API (genPK, sign, pkFromSig) for WOTS+. This way, we can
define other protocols based on the interface and instantiate them with
the approved WOTS parameter sets.

I duplicated most of the docs.