ec: replace mp_mod_inv with recip #100
This replaces the `mp_mod_inv` function with `recip`. The latter uses the
`Field` instance for `Z p` and is thus safer than `mp_mod_inv`, as it requires
that `p` is prime.

This is the only real call site of `mp_mod_inv`, so we can remove `mp_mod_inv`
in a follow-up commit.