utils: remove mp_mod_inv #100
`mp_mod_inv` has an incorrect type signature, as it is only valid when the
modulus is prime, but the type does not reflect this. Moreover, there is now a
correct version of this in the Cryptol prelude in the form of `recip : {p}
(prime p) => Z p -> Z p`, so there is really no need to define `mp_mod_inv`
separately anymore. Let's just remove `mp_mod_inv`.