Simon: Improve Specification
The Simon specification in the cryptol-specs repository was written before a style was codified.
This update brings the Simon specification up to date with the codified style.
This update also rewrites the Simon specification as a parameterized module.
The former specification was written before the Cryptol 3 module system was introduced.