ChaChaPolyCryptolIETF: Return an Option in AEAD_CHACHA20_POLY1305_DECRYPT #239
Instead of returning `([m][8], Bit)`, where the `Bit` indicates whether the
result is valid or not, we now make the `AEAD_CHACHA20_POLY1305_DECRYPT`
function return an `Option`. We also introduce an `optFold` function to
`Common::OptionUtils` to factor out some repeatedly used code needed to
post-process the decrypted results.

This checks off one box in #239.