ecdsa: don't mark checkSignature as a property #334
`checkSignature` would be run with `:check` (since it was marked as a
`property`) but then fail. `checkSignature` isn't meant to be used as a
standalone property anyway, but rather as a building block for constructing
other properties (e.g., `sampleVector`). Let's remove the `property` keyword to
reflect this and fix `:check`.

Fixes #334.